Work on splitting ExtractNestedTyping into 3 separate mods
